unify checks (#91)
