+++ tests/classification/test_accuracy.py
@@ -29,8 +29,8 @@ from tests.classification.inputs import _input_multilabel_multidim_prob as _inpu
 from tests.classification.inputs import _input_multilabel_prob as _input_mlb_prob
 from tests.helpers.testers import THRESHOLD, MetricTester
 from torchmetrics import Accuracy
-from torchmetrics.classification.checks import _input_format_classification
 from torchmetrics.functional import accuracy
+from torchmetrics.utilities.checks import _input_format_classification
 from torchmetrics.utilities.enums import DataType
 
 torch.manual_seed(42)
