+++ tests/classification/test_hamming_distance.py
@@ -26,8 +26,8 @@ from tests.classification.inputs import _input_multilabel_multidim_prob as _inpu
 from tests.classification.inputs import _input_multilabel_prob as _input_mlb_prob
 from tests.helpers.testers import THRESHOLD, MetricTester
 from torchmetrics import HammingDistance
-from torchmetrics.classification.checks import _input_format_classification
 from torchmetrics.functional import hamming_distance
+from torchmetrics.utilities.checks import _input_format_classification
 
 torch.manual_seed(42)
