+++ tests/classification/test_inputs.py
@@ -27,7 +27,7 @@ from tests.classification.inputs import _input_multilabel_multidim as _mlmd
 from tests.classification.inputs import _input_multilabel_multidim_prob as _mlmd_prob
 from tests.classification.inputs import _input_multilabel_prob as _ml_prob
 from tests.helpers.testers import BATCH_SIZE, EXTRA_DIM, NUM_BATCHES, NUM_CLASSES, THRESHOLD
-from torchmetrics.classification.checks import _input_format_classification
+from torchmetrics.utilities.checks import _input_format_classification
 from torchmetrics.utilities.data import select_topk, to_onehot
 from torchmetrics.utilities.enums import DataType
