+++ tests/classification/test_stat_scores.py
@@ -27,8 +27,8 @@ from tests.classification.inputs import _input_multilabel as _input_mcls
 from tests.classification.inputs import _input_multilabel_prob as _input_mlb_prob
 from tests.helpers.testers import NUM_CLASSES, THRESHOLD, MetricTester
 from torchmetrics import StatScores
-from torchmetrics.classification.checks import _input_format_classification
 from torchmetrics.functional import stat_scores
+from torchmetrics.utilities.checks import _input_format_classification
 
 torch.manual_seed(42)
