+++ torchmetrics/functional/classification/auroc.py
@@ -16,9 +16,9 @@ from typing import Optional, Sequence, Tuple
 
 import torch
 
-from torchmetrics.classification.checks import _input_format_classification
 from torchmetrics.functional.classification.auc import auc
 from torchmetrics.functional.classification.roc import roc
+from torchmetrics.utilities.checks import _input_format_classification
 from torchmetrics.utilities.enums import AverageMethod, DataType
